Rework traps API and add a few more tests
Changes:

- eliminates IPCZ_TRAP_CONDITION_DESTROYED. Instead we introduce
a IPCZ_DESTROY_TRAP_BLOCKING flag which allows a DestroyTrap() caller to
specify that the call must block until no more event handler invocations
are running for the trap.

- decouples a trap's lifetime from its portal's lifetime, so the
application's responsibility for managing trap lifetime is congruent
with other types of ipcz objects.

- context is a uint64_t rather than a uintptr_t for consistent width
  under any bitness.

+++ src/core/portal.h
@@ -69,7 +69,7 @@ class Portal : public mem::RefCounted {
 
   IpczResult CreateTrap(const IpczTrapConditions& conditions,
                         IpczTrapEventHandler handler,
-                        uintptr_t context,
+                        uint64_t context,
                         IpczHandle& trap);
   IpczResult ArmTrap(IpczHandle trap,
                      IpczTrapConditionFlags* satisfied_condition_flags,
